Add files via upload
In this Project we use the logistic regression and by using this logistic regression applies on the dataset after Encoding the data and removing the outliers so that our prediction is much accurate. We predict that which customer is absolutely can gives back the bank loan on time, bank give them loan on  the basis of the capability of that customer that how much they can pay back.